requirements.txt
durable-rules==2.0.28
jupyterlab==4.0.8
nltk==3.8.1
